// src/app/models/planets.model.ts
import { Films } from './films.modul';
import { People } from './people.model';

export interface Planets {
    name: string;
    rotation_period: number;
    orbital_period: number;
    diameter: number;
    climate: string;
    gravity: string;
    terrain: string;
    surface_water: number;
    population: number;
    residents?: People;
    films?: Films[];
}